implemented the idea of recognizing specifiers, wrote the %d specifier

+++ headers/myPrintf.hpp
@@ -1,6 +1,6 @@
 #ifndef MYPRINTF_HPP
 #define MYPRINTF_HPP
 
-extern "C" int myPrintf(const char* fmt, ...);
+extern "C" long myPrintf(const char* fmt, ...);
 
 #endif
